renaming ethRPC
